agent-context/context-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
python-dotenv>=1.0.0
# Document ingestion
unstructured[pdf,docx,html]>=0.21.0
# DB clients
psycopg2-binary>=2.9.9
